Stable version: no grab interactable
This version is for trying the best way to build the grab function.

The grab interactable has the problem that it somehow can change the light of the project, turning almost off the Passtrough material used for the shape of the objects. Next iteration will have the grab interaction working